Updated Deck and cards, Created driver program for Deck
Driver program for deck is the folder.
I setup the Deck so it initializes the cards from a file containing the
Type of card, Name and Description of the card.
I'm currently just using a delimiter to read in the information from the
file. Not a very dynamic since I tell it how many cards are in the file.

We need to use a STL data struct since this will be easier I think to
deal with the cards, otherwise we will have to create Methods to add,
remove, place card last etc...