Updating application core
I have reworked the application from the ground up. It is now building
and running, and should be moving through the first menus, and setting
up the game data. I consolidated the .h and .cpp files to make the
project easier to navigate. I tried to commet heavily, however in the
interest of time, some things are still uncommeted. I will write up a
change log in the near future and detail the specific functionality of
the application to make unit testing easier.