Updating the project to the current working version
I added all of the new methods and data, as well as the updated MVC. I
also cleaned up the repository a little bit. There is a folder labelled
“VS Project” that has the project for VS 2012 if anyone wants to just
copy that folder out and run it. I know this is clunky, but we are
getting really close to a finished product here.

Please please please please run test cases on some of the major methods
in the project ASAP, this is our last week!!!